move non fs helpers into new utils mod
