+++ crates/tauri-bundler/src/bundle/platform.rs
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-License-Identifier: MIT
 
-use super::common::CommandExt;
+use crate::utils::CommandExt;
 use std::process::Command;
 
 // Copyright 2019-2024 Tauri Programme within The Commons Conservancy
